refactor: common UI 파일 정리 완료 (#99)
