+++ apps/web/src/app/(with-global-layout)/auth/signup/page.tsx
@@ -3,11 +3,11 @@ import { useEffect, useState } from 'react';
 import { useSearchParams } from 'next/navigation';
 import { useRouter } from 'next/navigation';
 import { getAuthInfo, socialSignin, upsertAuthInfo, getExistsUser, getAuthLogout } from 'src/entities/auth/supabase';
-import { AuthCard } from 'src/components/auth/AuthCard';
 import { useUserStore } from 'src/entities/auth/stores/UserStore';
-import { LoadingSpinner } from 'src/components/auth/LoadingSpinner';
+// import { LoadingSpinner } from 'src/components/auth/LoadingSpinner';
 import { toast } from '@repo/ui/components/ui/sonner';
 import type { Role, Provider } from '../../../../entities/auth/types';
+import { AuthCard } from 'src/features/auth/AuthCard';
 
 export default function SignupPage() {
   const searchParams = useSearchParams();
@@ -99,7 +99,8 @@ export default function SignupPage() {
           <AuthCard title="회원가입" role={role} onTabChange={handleTabChange} onSocialSignin={handleSocialSignin} />
         )
       ) : (
-        <LoadingSpinner size={48} color="#8E74F9" />
+        // <LoadingSpinner size={48} color="#8E74F9" />
+        <div>로딩 중</div>
       )}
     </>
   );
