--- apps/web/src/shared/PageTitle.tsx
@@ -1,10 +1,0 @@
-interface PageTitleProps {
-  children: React.ReactNode;
-  className?: string;
-}
-
-const PageTitle = ({ children, className = '' }: PageTitleProps) => {
-  return <h2 className={`text-xl font-bold text-center ${className}`}>{children}</h2>;
-};
-
-export default PageTitle;
